requirements.txt
Django==5.2.1
djangorestframework>=3.14.0
psycopg2-binary>=2.9.6
python-dotenv==1.0.0   
urllib3==2.0.3       
django-environ>=0.11.2 